SalaryDAO 인터페이스 추가, SalaryMapper NAMESPACE 변경
